requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.21.0
requests>=2.31.0
pymupdf>=1.23.0
pyyaml>=6.0
watchdog>=3.0.0
ollama>=0.1.7
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.2.1
